pyproject.toml
[project]
name = "word-puzzle"
version = "0.1.0"
description = "A word matching app using API"
authors = [
    {name = "Matt Hemstock"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "email-validator (>=2.0.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
